Pull OnGenerate member to base class and introduce abstract GetCodeGenerator() method
