+++ src/ApiClientCodeGen.VSMac/CustomTools/BaseSingleFileCustomTool.cs
@@ -1,4 +1,6 @@
+using System.IO;
 using System.Threading.Tasks;
+using ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Core.Generators;
 using MonoDevelop.Core;
 using MonoDevelop.Ide.CustomTools;
 using MonoDevelop.Projects;
@@ -7,15 +9,25 @@ namespace ApiClientCodeGen.VSMac.CustomTools
 {
     public abstract class BaseSingleFileCustomTool : ISingleFileCustomTool
     {
-        public Task Generate(ProgressMonitor monitor, ProjectFile file, SingleFileCustomToolResult result)
+        public async Task Generate(ProgressMonitor monitor, ProjectFile file, SingleFileCustomToolResult result)
         {
             Bootstrapper.Initialize();
-            return OnGenerate(monitor, file, result);
+            
+            var path = file.FilePath.ChangeExtension(".cs");
+            result.GeneratedFilePath = path;
+
+            var customToolNamespace = file.CustomToolNamespace;
+            if (string.IsNullOrWhiteSpace(customToolNamespace))
+                customToolNamespace = CustomToolService.GetFileNamespace(file, path);
+
+            var generator = GetCodeGenerator(path, customToolNamespace);
+            var progressReporter = new ProgressReporter(monitor);
+            var contents = await Task.Run(() => generator.GenerateCode(progressReporter));
+            await Task.Run(() => File.WriteAllText(path, contents));
         }
 
-        protected abstract Task OnGenerate(
-            ProgressMonitor monitor,
-            ProjectFile file,
-            SingleFileCustomToolResult result);
+        protected abstract ICodeGenerator GetCodeGenerator(
+            string swaggerFile,
+            string customToolNamespace);
     }
 }
